Implement / move remaining subscriber methods for region selector
Implements and moves the missing abstract methods from the previous
commit, so that all buttons (which have intended implementation) work
correctly.
